get the edges right
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@13874 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ blind/plotimage.h
@@ -10,7 +10,7 @@ struct plotimage_args {
 	// FIXME -- alpha?
 
 	sip_t* wcs;
-	int gridsize;
+	double gridsize;
 
 	unsigned char* img;
 	int W;
